update: make extract agnostic about artifact types again
