+++ internal/utils/extract.go
@@ -24,13 +24,11 @@ import (
 	"os"
 	"path/filepath"
 	"strings"
-
-	"github.com/falcosecurity/falcoctl/pkg/oci"
 )
 
 // ExtractTarGz extracts a *.tar.gz compressed archive and moves its content to destDir.
 // Returns a slice containing the full path of the extracted files.
-func ExtractTarGz(gzipStream io.Reader, destDir string, artifactType oci.ArtifactType, stripPathComponents int) ([]string, error) {
+func ExtractTarGz(gzipStream io.Reader, destDir string, stripPathComponents int) ([]string, error) {
 	var files []string
 
 	uncompressedStream, err := gzip.NewReader(gzipStream)
@@ -59,16 +57,11 @@ func ExtractTarGz(gzipStream io.Reader, destDir string, artifactType oci.Artifac
 
 		switch header.Typeflag {
 		case tar.TypeDir:
-			if artifactType == oci.Plugin || artifactType == oci.Rulesfile {
-				return nil, fmt.Errorf("unexepected dir inside the archive, "+
-					"expected to find only files without any tree structure for %q artifacts", artifactType.String())
-			} else if artifactType == oci.Asset {
-				d := filepath.Join(destDir, strippedName)
-				if err = os.Mkdir(filepath.Clean(d), 0o750); err != nil {
-					return nil, err
-				}
-				files = append(files, d)
+			d := filepath.Join(destDir, strippedName)
+			if err = os.Mkdir(filepath.Clean(d), 0o750); err != nil {
+				return nil, err
 			}
+			files = append(files, d)
 		case tar.TypeReg:
 			f := filepath.Join(destDir, strippedName)
 			outFile, err := os.Create(filepath.Clean(f))
@@ -104,30 +97,4 @@ func stripComponents(headerName string, stripComponents int) string {
 		return headerName
 	}
 	return filepath.Clean(strings.Join(names[stripComponents:], "/"))
-}
-
-func listHeaders(gzipStream io.Reader) ([]string, error) {
-	uncompressedStream, err := gzip.NewReader(gzipStream)
-	if err != nil {
-		return nil, err
-	}
-
-	tarReader := tar.NewReader(uncompressedStream)
-
-	var files []string
-	for {
-		header, err := tarReader.Next()
-
-		if errors.Is(err, io.EOF) {
-			break
-		}
-
-		if err != nil {
-			return nil, err
-		}
-
-		files = append(files, header.Name)
-	}
-
-	return files, nil
 }
